requirements.txt
streamlit
supabase
python-dotenv
pandas
requests
mistralai
openpyxl